+++ landingzones/tests/test_permissions_ajax.py
@@ -84,6 +84,12 @@ class TestZoneStatusRetrieveAjaxView(LandingzonesPermissionTestBase):
             method='POST',
         )
 
+    def test_post_block(self):
+        """Test POST with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200, method='POST')
+        self.assert_response(self.url, self.non_superusers, 403, method='POST')
+
     def test_post_read_only(self):
         """Test POST with site read-only mode"""
         self.set_site_read_only()
@@ -151,6 +157,12 @@ class TestZoneStatusInfoRetrieveAjaxView(LandingzonesPermissionTestBase):
         self.project.set_public()
         self.assert_response(self.url, self.no_role_users, 403)
 
+    def test_get_block(self):
+        """Test GET with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200)
+        self.assert_response(self.url, self.non_superusers, 403)
+
     def test_get_read_only(self):
         """Test GET with site read-only mode"""
         self.set_site_read_only()
@@ -198,29 +210,35 @@ class TestZoneChecksumStatusRetrieveAjaxView(LandingzonesPermissionTestBase):
             self.anonymous,
         ]
 
-    def test_get(self):
-        """Test ZoneStatusInfoRetrieveAjaxView GET"""
+    def test_post(self):
+        """Test ZoneStatusInfoRetrieveAjaxView POST"""
         self.assert_response(self.url, self.good_users, 200, method='POST')
         self.assert_response(self.url, self.bad_users, 403, method='POST')
         self.project.set_public()
         self.assert_response(self.url, self.no_role_users, 403, method='POST')
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
-    def test_get_anon(self):
-        """Test GET with anonymous access"""
+    def test_post_anon(self):
+        """Test POST with anonymous access"""
         self.project.set_public()
         self.assert_response(self.url, self.anonymous, 403, method='POST')
 
-    def test_get_archive(self):
-        """Test GET with archived project"""
+    def test_post_archive(self):
+        """Test POST with archived project"""
         self.project.set_archive()
         self.assert_response(self.url, self.good_users, 200, method='POST')
         self.assert_response(self.url, self.bad_users, 403, method='POST')
         self.project.set_public()
         self.assert_response(self.url, self.no_role_users, 403, method='POST')
 
-    def test_get_read_only(self):
-        """Test GET with site read-only mode"""
+    def test_post_block(self):
+        """Test POST with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200, method='POST')
+        self.assert_response(self.url, self.non_superusers, 403, method='POST')
+
+    def test_post_read_only(self):
+        """Test POST with site read-only mode"""
         self.set_site_read_only()
         self.assert_response(self.url, self.good_users, 200, method='POST')
         self.assert_response(self.url, self.bad_users, 403, method='POST')