+++ landingzones/tests/test_permissions_ajax_taskflow.py
@@ -87,6 +87,12 @@ class TestZoneIrodsListRetrieveAjaxView(
         self.assert_response(self.url, self.good_users, 200)
         self.assert_response(self.url, self.bad_users, 403)
 
+    def test_get_block(self):
+        """Test GET with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200)
+        self.assert_response(self.url, self.non_superusers, 403)
+
     def test_get_read_only(self):
         """Test GET with site read-only mode"""
         self.set_site_read_only()