+++ samplesheets/tests/test_permissions_ajax.py
@@ -115,6 +115,12 @@ class TestSheetContextAjaxView(SampleSheetsAjaxPermissionTestBase):
         )
         self.assert_response(self.url, self.anonymous, 403)
 
+    def test_get_block(self):
+        """Test GET with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200)
+        self.assert_response(self.url, self.non_superusers, 403)
+
     def test_get_read_only(self):
         """Test GET with site read-only mode"""
         self.set_site_read_only()
@@ -159,6 +165,12 @@ class TestStudyTablesAjaxView(SampleSheetsAjaxPermissionTestBase):
             self.url, [self.user_finder_cat, self.user_no_roles], 200
         )
         self.assert_response(self.url, self.anonymous, 403)
+
+    def test_get_block(self):
+        """Test GET with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200)
+        self.assert_response(self.url, self.non_superusers, 403)
 
     def test_get_read_only(self):
         """Test GET with site read-only mode"""
@@ -206,8 +218,19 @@ class TestStudyTablesAjaxView(SampleSheetsAjaxPermissionTestBase):
             self.url, self.no_role_users, 403, data=self.edit_data
         )
 
+    def test_get_edit_block(self):
+        """Test GET with edit mode and project access block"""
+        app_settings.set(
+            'samplesheets', 'allow_editing', True, project=self.project
+        )
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200, data=self.edit_data)
+        self.assert_response(
+            self.url, self.non_superusers, 403, data=self.edit_data
+        )
+
     def test_get_edit_read_only(self):
-        """Test GET with site read-only mode"""
+        """Test GET with edit mode and site read-only mode"""
         app_settings.set(
             'samplesheets', 'allow_editing', True, project=self.project
         )
@@ -262,6 +285,12 @@ class TestStudyLinksAjaxView(SampleSheetsAjaxPermissionTestBase):
         self.assert_response(self.url, self.good_users_read, 404)  # No plugin
         self.assert_response(self.url, self.bad_users_read, 403)
 
+    def test_get_block(self):
+        """Test GET with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 404)
+        self.assert_response(self.url, self.non_superusers, 403)
+
     def test_get_read_only(self):
         """Test GET with site read-only mode"""
         self.set_site_read_only()
@@ -306,6 +335,12 @@ class TestSheetWarningsAjaxView(SampleSheetsAjaxPermissionTestBase):
         )
         self.assert_response(self.url, self.anonymous, 403)
 
+    def test_get_block(self):
+        """Test GET with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200)
+        self.assert_response(self.url, self.non_superusers, 403)
+
     def test_get_read_only(self):
         """Test GET with site read-only mode"""
         self.set_site_read_only()
@@ -345,6 +380,12 @@ class TestSheetCellEditAjaxView(SampleSheetsAjaxPermissionTestBase):
         self.assert_response(self.url, self.non_superusers, 403, method='POST')
         self.project.set_public()
         self.assert_response(self.url, self.no_role_users, 403, method='POST')
+
+    def test_post_block(self):
+        """Test POST with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200, method='POST')
+        self.assert_response(self.url, self.non_superusers, 403, method='POST')
 
     def test_post_read_only(self):
         """Test POST with site read-only mode"""
@@ -386,6 +427,12 @@ class TestSheetRowInsertAjaxView(SampleSheetsAjaxPermissionTestBase):
         self.project.set_public()
         self.assert_response(self.url, self.no_role_users, 403, method='POST')
 
+    def test_post_block(self):
+        """Test POST with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200, method='POST')
+        self.assert_response(self.url, self.non_superusers, 403, method='POST')
+
     def test_post_read_only(self):
         """Test POST with site read-only mode"""
         self.set_site_read_only()
@@ -425,6 +472,12 @@ class TestSheetRowDeleteAjaxView(SampleSheetsAjaxPermissionTestBase):
         self.assert_response(self.url, self.non_superusers, 403, method='POST')
         self.project.set_public()
         self.assert_response(self.url, self.no_role_users, 403, method='POST')
+
+    def test_post_block(self):
+        """Test POST with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200, method='POST')
+        self.assert_response(self.url, self.non_superusers, 403, method='POST')
 
     def test_post_read_only(self):
         """Test POST with site read-only mode"""
@@ -466,6 +519,12 @@ class TestSheetVersionSaveAjaxView(SampleSheetsAjaxPermissionTestBase):
         self.project.set_public()
         self.assert_response(self.url, self.no_role_users, 403, method='POST')
 
+    def test_post_block(self):
+        """Test POST with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200, method='POST')
+        self.assert_response(self.url, self.non_superusers, 403, method='POST')
+
     def test_post_read_only(self):
         """Test POST with site read-only mode"""
         self.set_site_read_only()
@@ -505,6 +564,12 @@ class TestSheetEditFinishAjaxView(SampleSheetsAjaxPermissionTestBase):
         self.assert_response(self.url, self.non_superusers, 403, method='POST')
         self.project.set_public()
         self.assert_response(self.url, self.no_role_users, 403, method='POST')
+
+    def test_post_block(self):
+        """Test POST with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200, method='POST')
+        self.assert_response(self.url, self.non_superusers, 403, method='POST')
 
     def test_post_read_only(self):
         """Test POST with site read-only mode"""
@@ -548,6 +613,12 @@ class TestSheetEditConfigAjaxView(SampleSheetsAjaxPermissionTestBase):
         self.project.set_public()
         self.assert_response(self.url, self.no_role_users, 403, method='POST')
 
+    def test_post_block(self):
+        """Test POST with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 400, method='POST')
+        self.assert_response(self.url, self.non_superusers, 403, method='POST')
+
     def test_post_read_only(self):
         """Test POST with site read-only mode"""
         self.set_site_read_only()
@@ -590,6 +661,12 @@ class TestStudyDisplayConfigAjaxView(SampleSheetsAjaxPermissionTestBase):
         self.project.set_public()
         self.assert_response(self.url, self.user_guest, 400, method='POST')
         self.assert_response(self.url, self.anonymous, 403, method='POST')
+
+    def test_post_block(self):
+        """Test POST with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 400, method='POST')
+        self.assert_response(self.url, self.non_superusers, 403, method='POST')
 
     def test_post_read_only(self):
         """Test POST with site read-only mode"""
@@ -637,6 +714,12 @@ class TestSheetVersionCompareAjaxView(SampleSheetsAjaxPermissionTestBase):
         self.project.set_public()
         self.assert_response(self.url, [self.user_no_roles], 200)
         self.assert_response(self.url, [self.anonymous], 403)
+
+    def test_get_block(self):
+        """Test GET with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(self.url, self.superuser, 200)
+        self.assert_response(self.url, self.non_superusers, 403)
 
     def test_get_read_only(self):
         """Test GET with site read-only mode"""
@@ -727,6 +810,25 @@ class TestIrodsDataRequestCreateAjaxView(SampleSheetsAjaxPermissionTestBase):
         )
         self.assert_response(
             self.url, self.anonymous, 403, method='POST', data=self.post_data
+        )
+
+    def test_post_block(self):
+        """Test POST with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(
+            self.url,
+            self.superuser,
+            200,
+            method='POST',
+            data=self.post_data,
+            cleanup_method=self._cleanup,
+        )
+        self.assert_response(
+            self.url,
+            self.non_superusers,
+            403,
+            method='POST',
+            data=self.post_data,
         )
 
     def test_post_read_only(self):
@@ -854,6 +956,25 @@ class TestIrodsDataRequestDeleteAjaxView(
         )
         self.assert_response(
             self.url, self.anonymous, 403, method='POST', data=self.post_data
+        )
+
+    def test_post_block(self):
+        """Test POST with project access block"""
+        self.set_access_block(self.project)
+        self.assert_response(
+            self.url,
+            self.superuser,
+            200,
+            method='POST',
+            data=self.post_data,
+            cleanup_method=self._cleanup,
+        )
+        self.assert_response(
+            self.url,
+            self.non_superusers,
+            403,
+            method='POST',
+            data=self.post_data,
         )
 
     def test_post_read_only(self):