+++ CHANGELOG.rst
@@ -24,6 +24,10 @@ Changed
     - Move black configuration to ``pyproject.toml`` (#2187)
     - Reformat migrations (#2187)
     - Reformat JQuery (#2187)
+- **Landingzones**
+    - Update views for project access blocking support (#2233)
+- **Samplesheets**
+    - Update views for project access blocking support (#2233)
 
 
 v1.1.4 (2025-08-12)